Added Goal and speed stars obtention

+++ Assets/Kitchen/Elements/LevelManager/Scripts/LevelData.cs
@@ -21,7 +21,7 @@ namespace Kitchen
         [Range(1, 4)] public int minNumberOfPainKillers;
 
         [Header("LevelGoals")]
-        [Range(1, 200)] public float time =200f;
+        [Range(1, 200)] public int time =200;
         [Range(1, 200)] public int goal =5;
 
         [Header("Recipes")]
